+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [x] Add purplepag.es to sign in flow
 - [x] Include people with only a display_name in search
 - [x] Fix AUTH over multiplextr
+- [x] Remember whether messages/notifications have been read
 
 ## 0.2.24